+++ src/main/java/de/samply/exporter/ExporterController.java
@@ -134,6 +134,22 @@ public class ExporterController {
                 exporterDbService::fetchAllQueries);
     }
 
+    @GetMapping(value = ExporterConst.QUERY_EXECUTIONS, produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<String> fetchQueryExecutions(
+            @RequestParam(name = ExporterConst.PAGE, required = false) Integer page,
+            @RequestParam(name = ExporterConst.PAGE_SIZE, required = false) Integer pageSize) {
+        return convertToResponseEntity(page, pageSize, exporterDbService::fetchAllQueryExecutions,
+                exporterDbService::fetchAllQueryExecutions);
+    }
+
+    @GetMapping(value = ExporterConst.QUERY_EXECUTION_ERRORS, produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<String> fetchQueryExecutionErrors(
+            @RequestParam(name = ExporterConst.PAGE, required = false) Integer page,
+            @RequestParam(name = ExporterConst.PAGE_SIZE, required = false) Integer pageSize) {
+        return convertToResponseEntity(page, pageSize, exporterDbService::fetchAllQueryExecutionErrors,
+                exporterDbService::fetchAllQueryExecutionErrors);
+    }
+
     @GetMapping(value = ExporterConst.INQUIRY, produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<String> fetchInquiry(HttpServletRequest httpServletRequest,
                                                @RequestHeader(name = ExporterConst.IS_INTERNAL_REQUEST, required = false) Boolean isInternalRequest,
@@ -353,7 +369,8 @@ public class ExporterController {
             @RequestParam(name = ExporterConst.PAGE, required = false) Integer pageCounter,
             @RequestParam(name = ExporterConst.PAGE_SIZE, required = false) Integer pageSize,
             @RequestParam(name = ExporterConst.PIVOT_VALUE, required = false) String pivotValue,
-            @RequestParam(name = ExporterConst.MERGE_FILES, required = false) Boolean mergeFiles) {
+            @RequestParam(name = ExporterConst.MERGE_FILES, required = false) Boolean mergeFiles,
+            @RequestParam(name = ExporterConst.FILE_AS_PLAIN_TEXT_IN_BODY, required = false) Boolean fileAsPlainTextInBody) {
         Optional<QueryExecution> queryExecution = exporterDbService.fetchQueryExecution(
                 queryExecutionId);
         if (queryExecution.isPresent()) {
@@ -369,7 +386,7 @@ public class ExporterController {
                 case OK -> {
                     try {
                         yield fetchQueryExecutionFilesAndZipOrMergeIfNecessary(
-                                queryExecutionId, fileFilter, fileColumnPivot, pageCounter, pivotValue, pageSize, mergeFiles);
+                                queryExecutionId, fileFilter, fileColumnPivot, pageCounter, pivotValue, pageSize, mergeFiles, fileAsPlainTextInBody);
                     } catch (ExporterControllerException | ZipperException | ExplorerException |
                              PivotIdentifierException | IOException e) {
                         yield createInternalServerError(e);
@@ -389,7 +406,7 @@ public class ExporterController {
 
     private ResponseEntity<InputStreamResource> fetchQueryExecutionFilesAndZipOrMergeIfNecessary(
             Long queryExecutionFileId, String fileFilter, String fileColumnPivot, Integer pageCounter,
-            String pivotValue, Integer pageSize, Boolean mergeFiles)
+            String pivotValue, Integer pageSize, Boolean mergeFiles, Boolean fileAsPlainTextInBody)
             throws ExporterControllerException, ZipperException, IOException, ExplorerException, PivotIdentifierException {
         List<QueryExecutionFile> queryExecutionFiles = exporterDbService.fetchQueryExecutionFilesByQueryExecutionId(
                 queryExecutionFileId);
@@ -403,10 +420,10 @@ public class ExporterController {
         files = filterFiles(files, fileFilter);
         if (files.size() > 0) {
             if (files.size() == 1) {
-                return createResponseEntity(files.get(0), httpHeaders);
+                return createResponseEntity(files.get(0), httpHeaders, fileAsPlainTextInBody);
             } else if (mergeFiles != null && mergeFiles == true && filesMergerManager.isFileExtensionSupported(files)) {
                 Optional<Path> mergedFile = filesMergerManager.merge(files);
-                return (mergedFile.isPresent()) ? createResponseEntity(mergedFile.get(), httpHeaders) : ResponseEntity.notFound().build();
+                return (mergedFile.isPresent()) ? createResponseEntity(mergedFile.get(), httpHeaders, fileAsPlainTextInBody) : ResponseEntity.notFound().build();
             } else {
                 Pair<InputStreamResource, String> inputStreamResourceFilenamePair = zipper.zipFiles(files);
                 return createResponseEntity(inputStreamResourceFilenamePair.getFirst(),
@@ -417,9 +434,11 @@ public class ExporterController {
         }
     }
 
-    private ResponseEntity<InputStreamResource> createResponseEntity(Path path, HttpHeaders httpHeaders) throws FileNotFoundException {
-        return createResponseEntity(new InputStreamResource(new FileInputStream(path.toAbsolutePath().toString())),
-                fetchFilename(path.getFileName().toString()), httpHeaders);
+    private ResponseEntity<InputStreamResource> createResponseEntity(Path path, HttpHeaders httpHeaders, Boolean fileAsPlainTextInBody) throws FileNotFoundException {
+        InputStreamResource inputStreamResource = new InputStreamResource(new FileInputStream(path.toAbsolutePath().toString()));
+        return (fileAsPlainTextInBody != null && fileAsPlainTextInBody) ?
+                createResponseEntityAsPlainTextInBody(inputStreamResource, httpHeaders) :
+                createResponseEntity(inputStreamResource, fetchFilename(path.getFileName().toString()), httpHeaders);
     }
 
     private List<Path> convertToPath(List<QueryExecutionFile> queryExecutionFiles) {
@@ -508,6 +527,11 @@ public class ExporterController {
             InputStreamResource inputStreamResource, String filename, HttpHeaders httpHeaders) {
         return ResponseEntity.ok().headers(httpHeaders).header("Content-Disposition", "attachment; filename=" + filename)
                 .body(inputStreamResource);
+    }
+
+    private ResponseEntity<InputStreamResource> createResponseEntityAsPlainTextInBody(
+            InputStreamResource inputStreamResource, HttpHeaders httpHeaders) {
+        return ResponseEntity.ok().headers(httpHeaders).contentType(MediaType.TEXT_PLAIN).body(inputStreamResource);
     }
 
     @GetMapping(value = ExporterConst.RETRIEVE_QUERY, produces = MediaType.APPLICATION_NDJSON_VALUE)