+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0 - 2023-10-17]
+## [1.1.0 - 2023-10-20]
 ### Added
 - Java Opts
 - Divide Excel file in several files if number of rows is too large
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Files Merger for JSON and XML
 - Input and Output formats as array in web service
 - CQL for Row Data
+- File as plain text in body
+- Fetch query executions
+- Fetch query execution errors
 
 ### Changed
 - Move FHIR Packages to dktk-exporter